First Commit
This PPTM will make creating an inclusive game easy for DBI's HR. This project is currently in progress and will be completed December 6th.
The PPTM will do most of the work for the user. The user will open the PPTM, which will have shapes labeled in the order in which they must be filled out. Once Everything is filled out and the user is content with what they built, the PPTM will format and save a file on that PC for the user to distribute.
Once the final users are finished with the game, it will automatically email the creator the scores of whoever completed the game.
As for now, this project will not completely function as it is currently in the middle of production. There are about 8 more things I would like to add before this is handed in.